many changes

+++ src/server/routes/controllers/chroniclesController.js
@@ -6,7 +6,7 @@ import Chronicle from '../../models/Chronicle';
 // Get all chronicles
 router.get("/", async (req, res) => {
     try {
-        res.json(await Chronicle.find());
+        res.json(await Chronicle.find( { storyTeller: req.session.userId }));
     }
     catch (e) {
         console.error(e);
@@ -14,9 +14,10 @@ router.get("/", async (req, res) => {
     }
 });
 
+// Get by status
 router.get("/status/:status", async (req, res) => {
     try {
-        res.json(await Chronicle.find({ status: req.params.status }));
+        res.json(await Chronicle.find({ status: req.params.status, storyTeller: req.session.userId }));
     } catch (e) {
         console.error(e);
         res.status(500).json(e);
@@ -26,7 +27,7 @@ router.get("/status/:status", async (req, res) => {
 // Get single chronicle
 router.get("/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findOne({ _id: req.params.id }));
+        res.json(await Chronicle.findOne({ _id: req.params.id, storyTeller: req.session.userId }));
     }
     catch (e) {
         console.error(e);
@@ -37,6 +38,7 @@ router.get("/:id", async (req, res) => {
 // Create new chronicle
 router.post("/", async (req, res) => {
     try {
+        req.body.storyTeller = req.session.userId;
         let chronicle = new Chronicle(req.body);
         res.json(await chronicle.save());
     }
@@ -49,7 +51,7 @@ router.post("/", async (req, res) => {
 // Update chronicle
 router.put("/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id }, req.body));
+        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id, storyTeller: req.session.userId }, req.body));
     }
     catch (e) {
         console.error(e);
@@ -60,7 +62,7 @@ router.put("/:id", async (req, res) => {
 // Delete chronicle
 router.delete("/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findByIdAndRemove({ _id: req.params.id }));
+        res.json(await Chronicle.findByIdAndRemove({ _id: req.params.id, storyTeller: req.session.userId }));
     }
     catch (e) {
         console.error(e);
@@ -70,7 +72,7 @@ router.delete("/:id", async (req, res) => {
 
 router.put("/archive/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id }, { status: "Archive" }));
+        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id, storyTeller: req.session.userId }, { status: "Archive" }));
     }
     catch (e) {
         console.error(e);
@@ -80,7 +82,7 @@ router.put("/archive/:id", async (req, res) => {
 
 router.put("/activate/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id }, { status: "Live" }));
+        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id, storyTeller: req.session.userId }, { status: "Live" }));
     }
     catch (e) {
         console.error(e);
@@ -90,7 +92,7 @@ router.put("/activate/:id", async (req, res) => {
 
 router.put("/resume/:id", async (req, res) => {
     try {
-        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id }, { status: "Draft" }));
+        res.json(await Chronicle.findByIdAndUpdate({ _id: req.params.id, storyTeller: req.session.userId }, { status: "Draft" }));
     }
     catch (e) {
         console.error(e);
